Added README help for deployment error
There's an error in deployment that MIGHT be unique to my new corporate laptop setup -- I didn't get it on my home computer or on my old work laptop, at least -- but just in case it happens to other users I've documented a workaround.

It has to do with a change in the way clasp & Google handle the path to your local project.  It used to be that clasp stored an absolute path to the local project (this is what the rootDir property means in .clasp.json), but this meant that it was effectively impossible to manage the same deployment on two different computers.

So now rootDir can be a relative path.  That's a good thing, but for some reason clasp cannot resolve the relative path on my new laptop, and this breaks our deployment script.  It may just be something broken in my npm install -- not every part of npm installed successfully due to corporate security constraints -- but in case it happens to someone else, you can fix this by manually pasting in the absolute path into the rootDir entry in .clasp.json.