Source passes are run an extra time by Interpreter.fromSheet.
Interpreter.fromSheet runs the SOURCE_PASSES multiple times as each sheet in
the workbook is added. Theses passes are re-run from start to finish over and
over with more symbols each time. Ultimately the code should be refactored to
avoid redoing work that has already been done, but this involves a fair amount
of work.

However, as things stand now, one extra processing of the SOURCE_PASSES is done
because Interpreter.fromSheet calls Interpreter.addSheet, which applies the
above iterative algorithm to resolve all symbols, and then Interpreter.fromSheet
runs the SOURCE_PASSES one more time to produce the grammar. This last time
through the SOURCE_PASSES is not needed if we save the grammar from the previous
run by Interpreter.addSheet.

This commit modified Interpreter.addSheet to store the grammar in the Workbook
object. Now Interpreter.fromSheet just retrieves the grammar from the Workbook
instead of running the SOURCE_PASSES an extra time.