Fixing clasp file-not-found issue
We figured out what was causing the deployment problem mentioned in last week's commit: the behavior of clasp's rootDir changed (for the better) in version 3.x, but our dev dependency was stuck to 2.x (where the behavior of rootDir was rather wacky).

When I was clasp-cloning manually, I was getting the new behavior (because 3.0.4 was installed globally), but as a script it was running 2.4.2.

This commit updates the dev-dependency to the new-and-better version.

Also some minor cleanups from the last few merges:

  -- Got rid of a debug log we didn't need anymore.
  -- We separated out SYMBOL_PASSES because we needed to run those separately in one place, but this was still, slightly confusingly, included in GRAMMAR_PASSES.  That's okay but it's nice to think of these pass phases as independent rather than worry how they nest/overlap.  So now there are three phrases that don't nest/overlap:

    * SOURCE_PASSES deal with turning the grids into grammars
    * SYMBOL_PASSES deal with creating and resolving symbols
    * TAPE_PASSES deal with tape/vocab resolution and transformations that require that information.