Generalizing lengthRange calculation
(1) In order to make good decisions about pre-compiling joins, we need to know (very roughly) the cardinality that a result set might have -- e.g. is it null?  is it 1 or a small number?  is it potentially infinite?

This is similar to the calculation we have to make about possible length ranges for infinity protection -- in fact, we already partly have it, it's the "null" boolean of lengthRange, which answers "Is the cardinality of the result set going to be zero (so far as we can know now) or non-zero?"

This generalizes that to actual numbers -- makes "null" into "cardinality" and makes it a numerical value rather than a boolean.

(1b) This is more than just a length calculation now, so I've renamed LengthRange "TapeSize" and lengthRange "getTapeSize".

I've also factored it out into a new module because it'll be used in several places.  (In the future, we might want to calculate it once and keep it around.)

(2) Found some more PriorityUnion holdouts and got rid of them.