More source operand tests and some error message cleanup.
Changed references to "tape name" in some error messages to something more
meaningful to the user. Also, "plain header" is now 'ordinary header", and
"parameter" or "argument" is now "operand" or "content". I also added more
details to some error messages (the short message, in particular), such as the
operator name or cell text in question.

When special headers that should be unique appear multiple times in a row, all
but the first instance should be ignored. The AssociateHeaders.transform() code
contained a small bug where all but the last instance were ignored; this bug
has been fixed so that the first instance is the one not ignored.

We added a toString() method for the Cell class.

We added 30 tests to testSourceOperand, increasing the number from 8 to 38 in
total.