Identifying compilation targets
We want to find joins with particular properties -- single shared tape, small finite cardinality on the right side, etc. -- as targets for JIT compilation later.

Doing so was a bit tricky, because we have to modify the symbol table, and the details of how the symbol table is threaded through a pass, and at what scopes we can access it.  Doing this transformation using the Auto-Pass boilerplate was losing the newly-created symbols.  Doing it more manually, just as a Pass, worked.

We've still got a bug in recursive filtering -- where a filter modifies a filter.  I'll get to that next.