Fixes for nested JIT targets
The previous commit didn't correctly handle nested JIT compilation targets; i.e., when we've identified a compilation target, but then it's in turn a child of another compilation target.  This isn't an uncommon situation; it's what you get when you put two filters on an embed, like

X=, embed, equals subj, equals tense
    Stem,  1SG,         PAST

The reasons were fairly minor, there were some places we were returning the old grammar g rather than the new one, and also we were neglecting to tapify.

NB: We can't actually tapify right at this juncture, the env we have access to isn't the right kind of env.  We could refactor to make this right, but there's no point -- we already know what the tapeSet is going to be, it's exactly the same as the original grammar.  So we just use that instead.

After this, there remains a bigger bug, that as-written we're assigning all filters of the same embed and of the same tape to the same symbol!  E.g. something like the following results in BOTH of these being replaced by a new symbol $Stem_subj despite them having different denotations.

X=, embed, equals subj
  , Stem, 1SG
  , Stem, 2SG

That was just a brain lapse, but we do have to be cautious about the solution, because not every right-hand-side here is going to be an easy string literal like 1SG or 2SG.  It should be sufficient, though, to use the toStr() of each grammar -- those should uniquely identify the denotation (at least for the limited class of grammars that can pass through the target-identification step in the first place).  But first I wanted to get these fixes committed.