More error message cleanup.
I changed the term "identifier" to "symbol name" or "header name", as appropriate.

More renaming of "tape" to "header" in error messages.

Added more details in some error messages, such as the content in error (e.g.,
the header name or text in error).

Clarified some error messages.

Added a few more tests.

Added short message text in all expected error results that previously did not
have them in the test suites.

I left test 12c in testSourceCollection in a failing state because that test
should be flagging a wayward collection error but it is not. This is likely due
to a bug where we are failing to propagate the error. This bug needs to be
tracked down.

I also want to track down why tests 25a and 25b of testSourceReplace return
different error messages.

The code for executing the test: operator needs scrutiny. Something wonky is
going on with the construction of the test grammars... it appears that each
test is nested with the previous. One consequence of this is that, in tests 5
and 7 of testSourceTest, the "ill-formed unit test" message is attached only
to the first test, but actually applies to all of the tests. One is lead to
believe that these test inputs are incorrect when, in fact, the issue is that
they did not execute, so did not return any results.

Other problems exist in the testSourceTest tests too, pointing at more issues
with test execution and reporting of errors. The "ill-formed unit test" error in
tests 5 & 7 should be flagged on location 12,1 or 12,4, not 13,2. Tests 13a &
13b flag the 'or:' operator, but the 'table:' operator should be flagged too.
Test 18 should return multiple errors.