Fixed remaining JIT target identification issue
In the last commit, there was still an outstanding issue that two filters with different filter values (e.g. equals subj: 1sg, vs equals subj: 2sg) would both be extracted into a new symbol with the same name.  This leads to incorrect answers, obviously, since you'd get (say) 2sg subjects even on the line for 1sg.

We hypothesized that the toStr() form of the grammar -- basically an s-expr -- would be sufficient to identify them uniquely and could serve as the name of the symbol for this purpose.  (It's not a valid identifier, but that's okay -- at this point it doesn't matter, and in fact it *shouldn't* be a valid identifier so that there's no chance of it shadowing a user-defined symbol.)

This seems to work fine.  NB that it's probably not true *in general*, like a given grammar G could have context-sensitive behavior that causes it to evaluate differently in different contexts.  (E.g., our recursion protection that caps off recursion at 4 means that G in one context vs. another might have different results.  Or if you put G inside a SingleTapeGrammar.)  But for the limited class of grammars we're actually dealing with here, the toStr() form should uniquely identify the set out outputs in the way we need it to.