Accumulative vocab collection
This is a test of a collectVocab function that only keeps one Set<string> per tape.

Stepping back, calculateTapes does three things:

    (1) figure out each grammar node's tapes
    (2) figure out whether the vocab of a node can be atomic or not
    (3) figure out what that vocab is

(2 and 3 aren't necessarily the same atomicity and vocab as the whole grammar, mind.  Just the subpart of the calculation that's valid for that node in the tree.)

I was hoping to try an algorithm that only did (3), because doing (2) in this manner is a bit tricky.  However, there were some places where I couldn't get ahold of (2) where I needed it -- something was unresolved, and I didn't want to wade back into the resolution algorithm to figure out why, not right now.  So I'm trying (2+3) simultaneously.

It seems to work so far, although I haven't rigorously tested it.  I haven't done any of the dicey bits about Embed, Match, etc., however, just the basics.