Much more work on testing the test: and testnot: operators.
"Unique" is now disallowed for the testnot operator.

Comments are now permitted in the unit tests (test & testnot).

Clarified more error messages.

We now pass env down in more places to facilitate debugging.

We now only return "Succeed" in the messages if there were no errors.

Reorganized the tests and added many more tests in testSourceTest so that now
positive tests (test:) are paired with corresponding negative tests (testnot:),
when appropriate.

Sorted out the generation of error messages for no matching outputs and extra
outputs, which wasn't working correctly in the context of uniqueness tests.

There still exist some issues with the test: and testnot: operators. The main
one is that the "Ill-formed unit test" messages for unknown headers should
appear on the test: or testnot: line attached to the header cell. A means to
pass the list of headers and their cell positions to ExecuteTests is needed,
maybe a TestSequenceGrammar or TestCollectionGrammar. This could be part of the
exercise to unwrap the nesting of tests within each other too. Also, currently
unknown headers with "unique" are not flagged as errors. The fix for flagging
the correct cell in the ill-formed test should test the "unique" headers too.

Tests 13a & 13b, involving or: operator under test: or testnot:, correctly flag
the or: operator with an error, but they should also flag the table: operator
that is present. The table: operator is correctly flagged as an error in tests
12a & 12b, which don't include the or: operator.

In our discussion about "unique", Pat suggested that we might want to change
"unique" to "output" and have two types of tests: inclusion tests, and
input/output tests.
Instead of the current:
	test:, text, unique gloss
the following syntax could be used for input/output tests:
    test:, input text, output gloss
We could also allow this to be written simply as:
    test:, text, output gloss