Fixed some issues in CheckStructuralParams with dropping error messages.
In commit 65fcac56894aa1eaf2d703f873f1cbbac07f32a1, we had left test 12c of
testSourceCollection in a failing state because it was failing to report some
errors; they were being dropped.

This commit tracked down this problem to CheckStructuralParams.handleOp
"throw"ing instead of "return"ing errors, which bypassed message collection
that was occurring higher up. We fixed this by having Msg.bind wrap the function
call in a try-catch that combines the message with those already collected.

This addressed the problem in test 12 of testSourceCollection - it's now
returning more messages. However, we realized that we still had the potential
to drop errors if we had nested errors involving a wayward assignment too
because the check for the wayward assignment error was checked ans collected
inside the function passed to bind before it call handleOp, so the try-catch is
still too far up. The real solution is not to use "throw", but to instead return
Msg<TST> from handleOp. Then the messages are collected properly. We left the
try-catch in bind as a fail-safe.

I added some tests to testSourceError (12a-12d) to explore nested errors, in
particular those involving wayward assignments and operators. Test 12d is the
key one demonstrating the phenomenon most completely. The errors are now caught.
However, test 12d reveals a new issue. It returns the message "'table' operator
requires grid, not 'text'", but the "text" cell should be a grid! It turns out
that it is seeing the "text" cell as a table operator, not a grid. This may have
something to do with a table op being inserted by the "InsertTables" pass for
the "noun=" assignment, which is later stripped as a wayward assignment, but the
inserted table op is left hanging around. More investigation is needed to track
down this problem, so I'm leaving test 12d in a failing state to remind us that
this needs fixing.