Further progress on vocab collection
(1) The previous commit only collected vocab from one tape at a time.  This worked as a proof-of-concept, but tapes' vocabularies are not independent objects, actions on one tape can affect another -- even in cases where that tape is hidden in the current scope.

To handle this, we have to collect all tapes' vocabs together, and the accumulator will have to be structurally complex -- ultimately quite similar to the old vocab object.  Tape names don't directly map to vocabs because there's no global one-to-one mapping of tape names and vocabs -- tape names are local!  Instead, tape names map to keys, and keys map to vocabs.  (Or rather, keys map to vocabs OR to other keys, to allow the kind of chained indirection.)

This is all being implemented alongside the old system -- the old vocabs objects are still there in grammars, for now I'm just adding additional fields.

(2) Our vocab testing still uses the old system (because we would have a sea of red otherwise).  So to test this, I stripped out the guts of testPassTapes and it's now testing this new system.

(2b) Testing vocab has become difficult ever since we went to "Tapes have no reality outside of their Cursors"; we've had to hack together some workarounds to get at the vocabs despite this.  Rather than rework those hacks to test the new vocab system, I just implemented a new recursive function, getVocab, that just grabs the first cursor it can find in its traversal with the right tapeName and returns its vocab.  This is good enough for the things we need to test.