Handling Matches in new vocab
(1) The new vocab system can now handle matches.  I ran into no fundamental problems; it's basically the same underlying idea as the previous one, so it should still work for all the tests.

(2) Unlike the old system, we're treating Match as always tokenizing.  In real grammars this will always be the case anyway, since they will contain dots.

(Also, I had realized there was some lack of clarity of what happens when we merge two tapes that are in different "modes", e.g. one's in "concat" mode and another is in "join".  What's the appropriate merge of those modes?  Is it symmetrical?  I don't think we thought about this in the old system, it was just whatever-the-system-happened-to-output.  We could ponder this but it would never matter, it's always going to be tokenized.  So let's just cut the Gordion know and say that Match, like Dot and Starts and some others, are always tokenizing.)

This changes a few test answers, since we had tests of atomic elements inside matches.  Now the resulting vocabs are tokenized.

(3) It's currently possible, in the old tape/vocab system, for Matches to reference tapes that don't exist.  (This won't happen in real grammars but it could in testing.)  So we have to check for whether the tapes exist before querying the VocabLibrary, otherwise we'll get tape not found errors.  When there's a bad tape we just don't do any merges, we collect vocab in the child as if the Match doesn't exist.