Finishing remaining grammars for get/collect vocab
(1) The remaining grammars were fairly smooth-sailing; we're green on testPassTapes now.  We even managed difficult tests like joining to a match outside of the scope of the cursor, first try.

A lot of the grammars we had previously had to consider in vocab we just don't, now, because they no longer exist at the stage that we resolve vocab: Starts, Ends, Contains, ReplaceBlock.

There might be a few grammars more in the rest of the tests we haven't handled; for now I was just trying to get testPassTapes green.

(2) PreTape was a bit annoying because we weren't actually keeping a vocab in it, in actual execution we just use the vocab of the output tape again.  But we do want to be able to test what that vocab is.  So now we're sticking a vocab in there just to test it.

(3) Added some testPassTapes for Cursors.  (We couldn't directly test those before, without a recursive function like the new getVocab().)