Hooking up new vocab to compilation
Slotting in the new vocab system in place of the old one was mostly uneventful; it revealed one or two typo-ish bugs in the new code but once those were fixed we were mostly green.

There is, however, one very strange change in the outputs -- a few repetition tests have additional, duplicated outputs now.  They all deal, in some way, with repeated repeats.  (Not necessarily the actual RepeatGrammar, just repeating something multiple times.)  E.g.:

    (t1:a+t2:b*)*.

That the outputs have multiple paths to them is unremarkable, many of these kinds of tests do.  For the most part, these don't appear in generation because of how outputs are unified, although there's a few that can't be unified and slip through the cracks.

We don't exactly consider this a bug, but it's very mysterious: why would a change to the *vocab* system change the output of these?  I looked at the vocabs for this and they seem to be the same as always.  In any case, we have bigger bugs to chase down right now, so this will have to wait.