Removing old vocab system from tapecalc
Removing vocab calculation from calculateTapes was a bit tricky because of how intertwined everything was; on several occasions I removed too much.  But there were no major issues.

There's one remaining red test -- an error that we previously caught and now don't.  It deals with absurd renames in embedded contexts -- e.g. where the rename is wrapped around a reference to its own symbol.  We're now not catching that error.  And also throw an exception because of a missing tape in the new vocab calc, but that's minor, we can avoid that.  The bigger thing is to understand why we're no longer catching that tape conflict.  (Maybe it's still unresolved at the point of tapecalc, and only later got resolved in resolveVocab?)  But in any case, I'm going to leave it unfixed for now while I clean up further.