Fixing rename+recursive embed bug
I actually fixed this in the previous commit, but forgot to add text to the commit message.  Consider this the commit message for both.

(1) We're no longer using vocab.js for anything, so I've removed it.  I also removed testPassVocab; this wasn't a general vocab test, just testing the complex objects that vocabs.js made.

(2) I figured out what was happening where we weren't catching the error when a rename happened in between a symbol assignment and a recursive reference to itself.  (I'm not sure what changed that we WERE catching it and then STOPPED, but I understand the problem and fixed the deeper issue.)

Backing up, the problem is when you have something like this:

    X = t1:hi + Rename(X, t1, t2)

This is paradoxical -- if X's tapes are {t1,t2}, then this rename is invalid because you can't rename to a tape that exists.  But X's tapes can't be {t1} because it clearly has a t2.  So we have to fix this up somehow, e.g. by removing the Rename.

What we were doing was a bit hacky because we didn't want to deal with the conceptual messiness of paradoxes yet, we just wanted to make sure this got an error and the Rename disappeared.  We were returning an empty tape literal when recursing, so that in recursion X has tapes {}. Then, because the Rename() tape function renames t1 to t2 inside of this object, the resulting tape literal is also {}.  This, then, triggers an error in tapeCalc-ing Renames -- or did, but now it doesn't.  (I think it's that there's no longer a last-minute tape resolution.)

This was rather indirect, however -- it didn't really have to do with t1 or t2, basically it was just an indirect way to detect that recursion is happening through a Rename and throwing an error.  And it leads to an odd error message that t1 isn't present, which isn't true in this instance, there's definitely a t1 in that rename.  (This is another instance of us having errors that reference something *we* do rather than the programmer does.)

Now I'm triggering the error not based on the absence of t1 in the grammar -- which was a bit of an artifact -- but the presence of t2.  That requires rewriting that Rename function so that it doesn't rename fromTape to toTape, but instead deletes fromTape and adds toTape.  This ensures that toTape is in there, even in recursive contexts, and then the outer Rename catches that t2 is already present.  And the error message makes more sense -- it refers to the presence of something that really exists in the code.

Ultimately, the problem here is that we don't have a true modern typechecking algorithm that can detect when two different lines of reasoning about the type of X assert incompatible assertions.  But we've never done that, and don't have plans to in the near future, because it's kind of overkill in a language where recursion is rare and explicit renaming is rare, too.  No one has actually ever run into this situation in a real project, so it's not a priority while we're still addressing big issues that do come up.