Using new getVocab for general vocab testing
(0) This isn't in this commit itself, but just for reference, I've confirmed that the new vocab system avoids the heap allocation error we were running into.

(1) Before this branch, we were testing the vocab of grammars in testGrammarUtils in an indirect way.  The tape opacity of Cursors made vocabs invisible where we needed them, and we hadn't gotten around to writing anything like getVocab.

So instead, we constructed a special grammar that would *generate* the vocabulary of a given tape t for a grammar G:

   t:. + Rep(G, 0, 0)

The zero-repetition of the grammar would ensure it didn't generate itself, while the dot would generate the single-character vocab.  (For these tests, we didn't want the atomicity optimization, or else every time we changed that optimization we would have to change the outputs of hundreds of tests!  But this gets handled as a side effect of dot -- when there's a dot, we don't do the atomicity optimization for that tape.)

Anyway, that was quite roundabout, and now we have getVocab.  So I rewrote testHasVocab to use it, along the lines of what we were doing in testPassTapes.  Then, I made testPassTapes use this new testHasVocab to avoid duplication of code.

(2a) Oops, I hadn't hooked up test and testnot to the new vocab system and they were throwing an exception.

(2b) We're still throwing exceptions is testSourceTest because we're not finding some symbols we need.  I haven't looked into it, but it's almost certainly due to the indirect way that test-running supplies its symbol table.  I'll fix that next.