Fixing vocab for tests 2
Just a small cleanup, but I forgot the commit explanation for the previous commit again.

The fix to get testSourceTest working again was straightforward but a little tedious.  The problem is that we weren't using the special SymbolEnv that the testing mechanism constructs, instead we were making a new symbol env.  At the point where we make this new env, we won't be going through a Qualified/SelectedGrammar anymore, and thus the symbol table was empty.

We weren't using it because, unlike other "staging" passes, we don't just use a SymbolEnv in resolveVocab, we use a subclass of it, ResolveVocabEnv, with extra information like a VocabLibrary and Stack.

Rather than construct a special ResolveVocabEnv that holds the special symbol table, I just realized that ResolveVocabEnv wasn't doing much.  (I had been threading vocabLib through the functions in various ways, meaning we didn't really need to thread it through via the env... and without this, ResolveVocabEnv really only provided a stack.  And getVocab wasn't even using this, only collectVocab was.)  So I threaded the stack through collectVocab, and made ResolveVocab just take a SymbolEnv.

With ResolveVocab now using the same SymbolEnv as every other staging pass, we no longer have to do anything special to get the tests to work again.