Fixed two issues with grids (header + content).
The first issue is that a grid with just a comment column was being treated as a
non-empty grid, i.e. it was being treated as if headers had been specified, with
no error being generated. I fixed this in CheckStructuralParams.handleOp by
allowing the headers for a grid to pass the non-empty test if at least one is not
a comment.

The second issue involved a grid with headers but no content below them. We got
different results if there was a commented column with content below the header
row (testSourceBasic 5e) than if there was no commented column
(testSourceBasic 5c-d). I solved this issue in AssociateHeaders.transform by
pretending there is a row of empty values in the grid if it only has headers.
Now test 5e gives the same results in generation as 5c-d; it doesn't generate
the warning because it in fact has a row of content values, courtesy of the
content value in the comment column.