Eliminating nested tests
We were creating TestGrammars and TestNotGrammars so that their structure was recursive.  If there were several rows of tests, each test's "child" (that grammar that it's testing) was itself the TestGrammar associated with the line above.  The motivation was just laziness, but the rational was basically: Since tests are supposed to be semantically transparent, it shouldn't matter whether its child is the previous test or the original grammar, it should end up testing that original grammar anyway...

However, you can end up getting strange interactions when there are programmer errors.  Also, when *we* have bugged code or bad choices elsewhere in the compiler, it makes it hard to reason about those.

So now, there's a TestBlockGrammar that stores the TestGrammars and TestNotGrammars corresponding to rows.  (If you have multiple TestBlockGrammars, they themselves still have a recursive structure, but that's "correct" in a way, all sequences of ops have that structure.)

I also changed the somewhat awkward interaction between test-tape issues (where the test references a tape that doesn't exist on the grammar) and test execution.  We used to execute the test anyway, but it's hard to reason about the results of ill-formed tests/testnots.  Now, instead of noting the header problem AND executing the test, we just don't execute the test and issue a warning on that test that it didn't run, so there's no confusion.

(However, in a future commit I'll probably move at least some of this error logic to calculateTapes.  We probably shouldn't be fussing about tape errors at execution time like this, we should do that during tapecalc.  But I didn't want to touch tapecalc too much yet, that's for another commit.)

I also haven't yet addressed the issue where we want test-tape errors to be flagged in the appropriate header cell.  That's also upcoming.