Making testblocks tape/vocab transparent
We had always said tests should be semantically transparent, that adding a test had no effects on the semantics of the symbol you added it to.  But this wasn't true: it changed the tapes/vocab, which in turn can change things like the interpretation of joins and dots!

So this fixes that; tests no longer contribute tapes/vocab to the grammar as a whole, just to the ephemeral JoinGrammar that we actually execute in executeTests.

I also added some tape/vocab tests to testSourceTest.  This actually wasn't straightforward, we had never tested tapes/vocab in Source tests before, and so had to add this functionality.

One snag in that functionality: in Source tests, by the time it comes to testHasVocab, prepareInterpereter already parsed and compiled the grammar according to the *original* opt, the one that optimizes atomicity.  So testHasVocab ended up ignoring the optimizeAtomicity: false that we were passing in; there's no mechanism to CHANGE an opt.

Either we re-parse all the sheets with this new tokenizing Opt, or we need to be able to swap out Opts after creation.  This latter is easy and quick but conceptually dicey, so we shouldn't do that in general, but *here* it's fine.  (Because of the recent vocab fixes, optimizeAtomicity now no longer has any effect during the prepareInterpreter steps. It's now only consulted later, during resolveVocab.)