Moving tape-error checking in unit tests
We were checking for tape-errors in unit tests right before running the test in executeTests. (And we were doing so repeatedly and redundantly, for each test line in a block, because only the test line actually had tape info, the TestBlock itself didn't).

This should really be caught and error-messaged in calculateTapes.  Now it is.