Putting test-tape error messages in proper cells
We had always flagged error messages about unknown tapes in tests on the "test" block cell (that is, the cell that says "test:" or "testnot:"), rather than in the header cell that caused the problem.  We didn't have the ability to do the latter, by the time we had tape information the location of the header cell was long gone.  (Grammars only have one location, and it's the location of their *content* cell, not their header cell.)

Now due to the changes in the previous commit, we have that information and can put it in the correct cell.  TestBlocks now keep a list of "dummy" grammars corresponding to their headers, that serve only to:

  (1) give TestBlock convenient access to information about what tapes are being tested
  (2) keep the location of the associated header cell, so that we know where to put error messages about tapes.

In fact, the last commit *should* have placed them in the right cell, but it didn't because we had a bit of a bug -- handleSingleTapes was dropping the position information.  (This position is usually redundant information, but it's not here -- this is the only memory of the header cell's location.)  So fixing that puts the errors in the right cells now.