Running tests "in parallel" for efficiency
We used to set up a new grammar and generate it for every single test.  This was inefficient for obvious reasons, and became even more inefficient with the recent changes to the vocab system -- we were running vocab collection over the whole grammar once for every test row!

Now, rather than join each test to the grammar and execute it, we take all the tests in a testblock, combine those into an alternation, and join that alternation to the grammar.  That way we only have to do the process once per testblock.

This, of course, "scrambles" the results so we don't know which results correspond to which test row.  To get around this, each test-row now contains a special literal $r:N, where N is the row on which it occurs.  (We don't need to qualify this with the sheet name, you'll never have $r from other sheets, because testblocks are now tape-transparent.)  That literal will show up in all results generated from that row, so we can use it to re-associate rows and results, and thus put success/failure messages in the right cells.